More options
Added a bunch of IF statements to lead through the dungeon, as well as some exceptions for handling invalid input and choices I don't like (like going back rather than forward, because that's a pain for now). Stopping until I learn functions because that'll apparently make my life a lot easier.