Function loop error
This is an example of an error I ran into when creating functions. When the user enters invalid input into the acc_var parameter, it directs the user to the acc_error function. At this point, if the user enters invalid input again, everything works as it should. But if the user enters valid input, it returns "None" for accomodation(acc_var) in the print on line 22.